+++ NEWS.md
@@ -10,6 +10,7 @@
 * In the case of long labels, attempt to auto-adapt text size in PCA coefficient and logFC plots
 * Add overview and crosslinks in the beginning of the reports
 * Include bar plot of significant features in pdf output
+* Allow iColPattern without Sample for PDTMT
 
 # einprot 0.6.7